Add files via upload
It's a Number Guessing Game using JAVA , in which user have to guess a number and the game indicates whether it is near the actual number or too high compare to it. Game at last also give the number of attempts user take to guess the correct number. 
In this Game I have used Swing to give it GUI Interface.